requirements.txt
certifi==2025.11.12
cfchecker==4.1.0
cftime==1.6.5
cfunits==3.3.7
charset-normalizer==3.4.4
future==1.0.0
idna==3.11
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
netCDF4==1.7.3
numpy==2.4.0
packaging==25.0
pandas==2.3.3
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
rdflib==7.5.0
requests==2.32.5
rich==14.2.0
six==1.17.0
tzdata==2025.3
urllib3==2.6.2
xarray==2025.12.0
